+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+## Apollo Client 3.4.11 (not yet released)
+
+### Bug Fixes
+
+- Fix [Vite](https://vitejs.dev) tree-shaking by calling the `checkDEV()` function (at least once) in the module that exports it, `@apollo/client/utilities/globals/index.ts`. <br/>
+  [@benjamn](https://github.com/benjamn) in [#8767](https://github.com/apollographql/apollo-client/pull/8767)
+
 ## Apollo Client 3.4.10
 
 ### Improvements